requirements.txt
# requirements.txt

openai
langchain
faiss-cpu
pandas
streamlit
python-dotenv
langchain-community
